Move common workflow functions into a workflow_common module.
